+++ src/Element/common.h
@@ -86,13 +86,13 @@ typedef struct s_element
 
 typedef struct s_range
 {
-	float	min;
-	float	max;
+	double	min;
+	double	max;
 }	t_range;
 
 typedef struct s_hit
 {
-	float		t;
+	double		t;
 	t_point		p;
 	t_vector3	normal;
 	t_color		color;
@@ -102,6 +102,6 @@ typedef struct s_hit
 extern void		init_element(void *elem);
 extern void		element_iter(t_list *list, enum e_element_func e);
 extern int		parse_rgb(struct s_rgb *rgb, char *curr);
-extern int		parse_vector3(t_vector3 *vector, char *curr, float scope);
+extern int		parse_vector3(t_vector3 *vector, char *curr, int scope);
 
 #endif
