+++ src/Element/util/element_util.h
@@ -18,17 +18,17 @@
 
 typedef struct s_abc
 {
-	float	a;
-	float	b;
-	float	c;
+	double	a;
+	double	b;
+	double	c;
 }	t_abc;
 
-extern t_point		get_ray_point(const t_ray *ray, float t);
+extern t_point		get_ray_point(const t_ray *ray, double t);
 extern t_vector3	get_face_normal(t_ray ray, t_vector3 normal);
-extern int			range_in(float t, t_range range);
-extern int			range_not_in(float t, t_range range);
-extern int			close_to_zero(float f);
-extern int			quadratic_formula_root(t_abc abc, t_range range, float *x);
+extern int			range_in(double t, t_range range);
+extern int			range_not_in(double t, t_range range);
+extern int			close_to_zero(double f);
+extern int			quadratic_formula_root(t_abc abc, t_range range, double *x);
 extern int			transform_element(
 						t_element *self,
 						const t_element *camera,
