+++ src/Element/util/range_not_in.c
@@ -12,7 +12,7 @@
 
 #include "element_util.h"
 
-inline int	range_not_in(float t, t_range range)
+inline int	range_not_in(double t, t_range range)
 {
 	return (t < range.min || t > range.max);
 }
